move hook to service
