+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 ## Unreleased
 
 ### Added
+
 - SC-6172: added hooks to check for unique and not disposable emails to adminUsers service
 
 ## [24.4.0] - 2020-8-31
